added joi, move auth to single auth model
