+++ packages/foodie-backend/package.json
@@ -27,7 +27,7 @@
     "start": "node ./bin/www.js",
     "server": "nodemon ./bin/www.js",
     "build": "npm run webpack & npm run tsc",
-    "serve": "concurrently --kill-others \"npm run tsc-watch\" \"npm run server\"",
+    "dev": "concurrently --kill-others \"npm run tsc-watch\" \"npm run server\"",
     "tsc": "tsc",
     "tsc-watch": "tsc --watch",
     "test": "jest",
@@ -54,6 +54,7 @@
     "formidable": "^1.2.1",
     "hbs": "^4.0.1",
     "helmet": "^3.15.0",
+    "joi": "^14.3.1",
     "jsonwebtoken": "^8.4.0",
     "mongodb": "^3.1.10",
     "multer": "^1.3.1",
