+++ packages/foodie-backend/src/app/auth/AuthActions.ts
@@ -1,7 +1,8 @@
 import { validate } from 'joi';
 import { IUser, UserData } from '../user/UserData';
-import { comparePassword, generateToken, ILogin } from './AuthData';
+import { comparePassword, generateToken, ILogin, encryptPassword } from './AuthData';
 import { LoginSchema } from './AuthSchema';
+import { RegisterSchema } from '../user/UserSchema';
 
 export async function Login(data: ILogin) {
 
@@ -12,9 +13,7 @@ export async function Login(data: ILogin) {
   const user = await UserData().findOne({ email });
 
   if (!user) {
-
-    throw new Error('user not found');
-
+    throw new Error('NOT FOUND');
   }
 
   const passwordCheck = await comparePassword(password, user);
@@ -27,7 +26,33 @@ export async function Login(data: ILogin) {
   delete user.password;
 
   return user;
+}
 
+export async function SignUpUser(data: IUser) {
+
+  const res = await validate(data, RegisterSchema);
+
+  const { email, password } = res;
+
+  const doesUserExist = await UserData().findOne({ email });
+
+  if (doesUserExist) {
+    throw new Error('user already exists');
+  }
+
+  const encryptedPassword = await encryptPassword(password);
+
+  res.password = encryptedPassword;
+
+  const result = await UserData().create(res);
+
+  const token = generateToken(result._id);
+
+  Object.assign(result, { token });
+
+  delete result.password;
+
+  return result;
 }
 
 export async function Logout() { }
