+++ packages/foodie-backend/src/app/auth/AuthRouter.ts
@@ -1,8 +1,9 @@
 import { Router } from 'express';
-import { Login } from './AuthController';
+import { Login, SignUpUser } from './AuthController';
 
 const router = Router();
 
 router.post('/login', Login);
+router.post('/register', SignUpUser);
 
 export default router;
