+++ packages/foodie-backend/src/app/user/UserRepository.ts
@@ -2,5 +2,3 @@ import { BaseRepository } from '../../repositories/BaseRepository';
 import { IUser } from './UserData';
 
 export default class UserRepository extends BaseRepository<IUser> { }
-
-// add custom data logic here to extend that o fht ebase repository
