+++ packages/foodie-backend/src/app/user/UserRouter.ts
@@ -1,12 +1,10 @@
 import { Router } from 'express';
-import { GetUser, GetUsers, SignUpUser } from './UserController';
+import { GetUser, GetUsers } from './UserController';
 
 const router = Router();
 
 router.get('/users', GetUsers);
 
 router.get('/user/:name/:id', GetUser);
-
-router.post('/user/register', SignUpUser);
 
 export default router;
